### 2025-01-22 #7706 @editorialbot
[PRE REVIEW]: Pydre: A Python package for driving simulation data reduction
**Submitting author:** @tkerwin (<a href="http://orcid.org/0000-0003-4560-3203">Thomas Kerwin</a>)
**Repository:** https://github.com/OSUDSL/pydre
**Branch with paper.md** (empty if default branch): joss
**Version:** v24.0.2
