### 2025-01-22 28263ba
add doi for "Simulator Data Reduction
