+++ joss/paper.bib
@@ -30,7 +30,8 @@
   booktitle = {Handbook of Driving Simulation for Engineering,  Medicine, and Psychology},
   author = {{Michelle L Reyes} and {John D Lee}},
   date = {2011},
-  publisher = {CRC Press}
+  publisher = {CRC Press},
+  doi = {10.1201/b10836-21}
 }
 
 @inproceedings{pawarDataReduction2020,
