### 2025-01-22 3b795e9
Support file list in project config
refactored some config stuff into Project
