### 2025-01-23 44a9e81
Strip path from string before parsing filename
