+++ pyproject.toml
@@ -8,7 +8,8 @@ dependencies = [
     "scipy",
     "icecream",
     "tqdm",
-    "loguru"
+    "loguru",
+    "jenkspy"
 ]
 requires-python = ">= 3.12"
 authors = [ {name = "Thomas Kerwin", email = "kerwin.6@osu.edu"} ]
