### 2025-01-24 #7706 @tkerwin
Perhaps users ajoer and fredshone would be interested.
