### 2025-01-24 0171913
make sure we don't modify product definition
use deepcopy
