### 2025-01-27 #7706 @arfon
👋 @ajoer @fredshone :wave: would you be willing to review this submission for JOSS? The submission under consideration is _Pydre: A Python package for driving simulation data reduction_

Many thanks
Arfon
