### 2025-01-27 6c2a63d
Initial Jenkspy function
function needs work (doesnt output binary column yet)
