### 2025-01-27 dee7df2
typo in changing to deep copy
