### 2025-01-28 3fc93cf
Added intermediate file
Intermediate csv file to inspect binary data
