### 2025-01-28 13471b7
fix error from not having a datafile at all
