### 2025-01-28 7914823
added biopticDipMeasure metric to calculate dip stats
