+++ project_files/btsHead.toml
@@ -1,16 +1,23 @@
+[config]
+datafiles = ["C:/Users/tchhajed/Documents/DSL/datplots_data/bioptics/*.dat"]
+
 [filters.fixHeadPitchData]
 function = "Jenks"
 oldCol = "HeadPitch"
 newCol = "hpBinary"
 
-[filters.CSV]
-function = "writeToCSV"
-outputDirectory = "/Users/saanazia/Documents/DSL/pydre/tests"
+[filters.numberBinaryBlocks]
+function = "numberBinaryBlocks"
+binary_column = "hpBinary"
+new_column = "DipRegions"
+only_on=0
+
+[metrics.HeadPitchStats]
+function = "biopticDipMeasure"
 
 [metrics."Max Head Pitch"]
 function = "colMax"
 var = "HeadPitch"
-
 
 [metrics."Min Head Pitch"]
 function = "colMin"
