### 2025-01-29 a6c8cfd
Fix up metadata for rect roi
