+++ project_files/r2dNonEvents.toml
@@ -1,0 +1,28 @@
+[filters.SimTimeFromDatTime]
+function = "SimTimeFromDatTime"
+
+[filters.FixReversedRoadLinearLand]
+function = "FixReversedRoadLinearLand"
+
+[filters.RegionNonEvent]
+function = "DesignateNonEventRegions"
+dataFile = 'C:\Users\Owner\Desktop\work\bama_work\internal_code\osu\merge_data\NO_EVENT_REGIONv1.csv'
+
+[rois.roi1]
+type = "column"
+columnname = "NonEventRegion"
+
+[metrics.R2DIDColumns]
+function = "R2DIDColumns"
+
+[metrics.NonEventRegionStart]
+function = "colMin"
+var = "XPos"
+
+[metrics.NonEventRegionEnd]
+function = "colMax"
+var = "XPos"
+
+[metrics.sdLP]
+function = "colSD"
+var = "RoadOffset"
