### 2025-02-05 0a1c4b5
Initial odd data filter
