### 2025-02-11 b6b48f9
Remove data outside/inside bounds!
It works! woohoo!
