### 2025-02-12 eeefc41
Fixed column name default and updated documentation
