### 2025-02-12 1a0e0cb
initial separate data function
- still have to test
