### 2025-02-18 a6b2c8d
Non-event lane violation metric - start of steering reversal
