### 2025-02-18 d2649be
Logger messages, Unhandled keywords in project files
