### 2025-02-19 a717364
time roi with specified metadata
