### 2025-02-20 f3cb8a2
Initial filterValuesBelow filter
should filter values below 1 m/s
