### 2025-02-20 8be47a4
fix column check in steering reversal metric
