### 2025-02-21 9b586fe
Updated filterValuesBelow
logic was incorrect. Still doesn't run.
