### 2025-02-21 248b215
fixed type errors for time roi
