### 2025-02-22 #7706 @arfon
@sgbaird @sumedhsakdeo 👋 would you be willing to review this submission for JOSS? The submission under consideration is Pydre: A Python package for driving simulation data reduction

Many thanks
Arfon
