### 2025-03-01 #7706 @arfon
@fxjung @xoolive @ixjlyons 👋 would you be willing to review this submission for JOSS? The submission under consideration is Pydre: A Python package for driving simulation data reduction

Based on your experience, (and related historical submission from @fxjung) we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
