### 2025-03-03 40d7193
update docs to remove json projects
