### 2025-03-04 #7870 @editorialbot
[REVIEW]: Pydre: A Python package for driving simulation data reduction
**Submitting author:** @tkerwin (<a href="http://orcid.org/0000-0003-4560-3203">Thomas Kerwin</a>)
**Repository:** https://github.com/OSUDSL/pydre
**Branch with paper.md** (empty if default branch): joss
**Version:** 25.2.1
**Archive:** 10.6084/m9.figshare.28765025.v1
