### 2025-03-04 #7870 @arfon
@fredshone, @xoolive, @fxjung – many thanks for agreeing to review this submission for JOSS! This is the review thread for the paper. All of our communications will happen here from now on.
