### 2025-03-04 87eea9b
Forgot to update version info in pyproject.toml
