+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pydre"
-version = "25"
+version = "25.0"
 readme = "readme.md"
 dependencies = [
     "polars",
