### 2025-03-05 f8ebe81
Moved bioptic specific project files to other repo
