+++ project_files/regular.toml
@@ -1,42 +1,0 @@
-[config]
-#Pick all data files
-datafiles = ["/Users/saanazia/Documents/DSL/bioptic/Data/*.dat"]
-ignore = [
-    "/Users/saanazia/Documents/DSL/bioptic/Data/Experimenter_9_Scenario 4_1727278726.dat",
-    "/Users/saanazia/Documents/DSL/bioptic/Data/Experimenter_6_Scenario 4_1723145764.dat"
-]
-
-
-[filters.fixHeadPitchData]
-function = "Jenks"
-oldCol = "HeadPitch"
-newCol = "hpBinary"
-
-[filters.removeStart]
-function = "filterValuesBelow"
-col = "Velocity"
-
-
-[filters.numberBinaryBlocks]
-function = "numberBinaryBlocks"
-binary_column = "hpBinary"
-new_column = "DipRegions"
-only_on=0
-
-[filters.rawData]
-function = "writeToCSV"
-outputDirectory = "/tmp/pydre"
-
-
-[metrics.HeadPitchStats]
-function = "biopticDipMeasure"
-
-[metrics."Max Head Pitch"]
-function = "colMax"
-var = "HeadPitch"
-
-[metrics."Min Head Pitch"]
-function = "colMin"
-var = "HeadPitch"
-
-
