+++ project_files/separate.toml
@@ -1,37 +1,0 @@
-[config]
-#Only file: Experimenter_9_Scenario 4_1727278726.dat
-datafiles = ['/Users/saanazia/Documents/DSL/bioptic/Data/Experimenter_9_Scenario 4_1727278726.dat']
-[filters.fixHeadPitchData]
-function = "Jenks"
-oldCol = "HeadPitch"
-newCol = "hpBinary"
-
-[filters.removeStart]
-function = "filterValuesBelow"
-col = "Velocity"
-
-[filters.numberBinaryBlocks]
-function = "numberBinaryBlocks"
-binary_column = "hpBinary"
-new_column = "DipRegions"
-only_on=0
-
-[filters.separateHP]
-function = "separateData"
-col = "HeadPitch"
-threshold = 1.0
-
-[filters.rawData]
-function = "writeToCSV"
-outputDirectory = "/tmp/pydre"
-
-[metrics.HeadPitchStats]
-function = "biopticDipMeasure"
-
-[metrics."Max Head Pitch"]
-function = "colMax"
-var = "HeadPitch"
-
-[metrics."Min Head Pitch"]
-function = "colMin"
-var = "HeadPitch"
