### 2025-03-05 5f0fcd1
Fix some type problems
also reformat
