### 2025-03-06 #7870 @fredshone
@tkerwin nice project and paper!

Some pretty minor changes please:

- Add installation instructions (ideally including venv for example) - suggest aim these at someone new to python.
- Your dev install instructions are incomplete (I think). After install (with uv) i have to use `uv run python -m src.pydre.run...` for example, tried on UNIX and Windows 11, basically `pydre` isn't executable? Perhaps this works with release installs?
- Please add docs for running tests and a bit more for community guidelines, [eg](https://github.com/arfc/osier/blob/main/docs/source/contrib.md).
- You have broken tests i think...
- Please add test coverage report, for example add pytest-cov to install and I suggest setting a fail under, [eg](https://github.com/arfc/osier/blob/3c986d6737b7335a7c48dd9a7a9f3bd2458bbc14/pyproject.toml#L78).

Minor thoughts:
- Your dev install docs use Rye but i think this is intended to be replaced by uv now?
- Your example usage in docs is a bit clunky to run (i had to work out the paths).

Otherwise really nice project! very clear use case and paper.
