### 2025-03-06 3bd8ee8
Added coverage module and contrib guidelines
