+++ pyproject.toml
@@ -43,6 +43,7 @@ dev-dependencies = [
     "black",
     "mypy",
     "pytest>=8.3.2",
+    "pytest-cov",
     "pytest-datafiles",
     "griffe==1.2.0"
 ]
