### 2025-03-06 #7870 @tkerwin
A general comment: I released a new version (25.0) two days ago. That version (hopefully) has fewer bugs and better documentation.

@fredshone Thank you very much for your comments.

I am working on installation instructions, docs for running tests and contribution guidelines currently. Thank you for the link to the excellent example for contribution guidelines. I expect to push those documentation updates soon and will update here when that it done.  I can add a test coverage report as well, although we not yet close to full coverage

Although uv is recommended for new projects, rye is still being actively maintained, so I don't think there's a need to shift off of it for now. However, the published package should work correctly with uv. I just ran the following steps on my machine with the expected results:

```
mkdir pydre-test
cd pydre-test
uv init
uv add pydre
uv sync
uv run python -m pydre.run
```

If you get an error with those steps, can you please make an [issue](https://github.com/OSUDSL/pydre/issues)? I haven't tried running pydre from the working development directory with uv, so maybe that's the use case you are getting the error in?

If you can also give me a little more detail about the broken tests, I would greatly appreciate it.

Thanks again for your review.
