### 2025-03-06 OSUDSL/pydre#1 @fredshone
Failing tests due to missing test fixtures
after:
```
git clone git clone git@github.com:OSUDSL/pydre.git
cd pydre
uv venv
uv sync
uv run pytest
```

some tests failing to find fixtures:

```
tests/test_metrics_common.py .......                                     [ 63%]
tests/test_project.py EE.E                                               [100%]

==================================== ERRORS ====================================
___________________ ERROR at setup of test_project_loadjson ____________________
file /home/fred/Projects/pydre/tests/test_project.py, line 9
  @pytest.mark.datafiles(FIXTURE_DIR / "good_projectfiles" / "test1_pf.json")
  def test_project_loadjson(datafiles):
E       fixture 'datafiles' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, doctest_namespace, monkeypatch, propagate_logs, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, reportlog, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

/home/fred/Projects/pydre/tests/test_project.py:9
___________________ ERROR at setup of test_project_loadtoml ____________________
file /home/fred/Projects/pydre/tests/test_project.py, line 15
  @pytest.mark.datafiles(FIXTURE_DIR / "good_projectfiles" / "test1_pf.toml")
  def test_project_loadtoml(datafiles):
E       fixture 'datafiles' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, doctest_namespace, monkeypatch, propagate_logs, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, reportlog, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.

/home/fred/Projects/pydre/tests/test_project.py:15
___________________ ERROR at setup of test_project_projequiv ___________________
file /home/fred/Projects/pydre/tests/test_project.py, line 26
  @pytest.mark.datafiles(
      FIXTURE_DIR / "good_projectfiles" / "test1_pf.json",
      FIXTURE_DIR / "good_projectfiles" / "test1_pf.toml",
  )
  def test_project_projequiv(datafiles):
E       fixture 'datafiles' not found
>       available fixtures: cache, capfd, capfdbinary, caplog, capsys, capsysbinary, doctest_namespace, monkeypatch, propagate_logs, pytestconfig, record_property, record_testsuite_property, record_xml_attribute, recwarn, reportlog, tmp_path, tmp_path_factory, tmpdir, tmpdir_factory
>       use 'pytest --fixtures [testpath]' for help on them.
```

i'm not familiar with [this use of pytest.mark](https://github.com/OSUDSL/pydre/blob/afd82bc57ebd786533b0f733b1c82357fa450868/tests/test_project.py#L9), perhaps it should be [adding fixtures from conftest](https://docs.pytest.org/en/stable/how-to/fixtures.html#usefixtures)?
