### 2025-03-06 #7870 @fredshone
Sounds good.

re coverage - no need to go for 100% - but just need to check there is a reasonable amount. I completely appreciate getting coverage for data pipelines is painful. So no stress.

re rye versus uv, fine by me (I'm not familiar with either in any case), those uv steps are close to what I did - I will await your updated docs.

re tests: [issue here](https://github.com/OSUDSL/pydre/issues/1), but might simply be my install.
