### 2025-03-06 50e5dd4
Add dev dependencies for uv compatibility
