+++ pyproject.toml
@@ -33,6 +33,20 @@ pydre-run = "pydre:run"
 Documentation = "https://osudsl.github.io/pydre/"
 Repository = "https://github.com/OSUDSL/pydre"
 
+[dependency-groups]
+dev = [
+    "ptpython",
+    "mkdocs>=1.6.0",
+    "mkdocs-material",
+    "mkdocstrings[python]",
+    "black",
+    "mypy",
+    "pytest>=8.3.2",
+    "pytest-cov",
+    "pytest-datafiles",
+    "griffe==1.2.0"
+]
+
 [tool.rye]
 managed = true
 dev-dependencies = [
