### 2025-03-06 c32d483
default to display test coverage
