### 2025-03-06 OSUDSL/pydre#1 @tkerwin
I tried to reproduce your issue and found out that my `pyproject.toml` file did not have the information to properly install the dev requirements for uv. Because of this, I couldn't run `uv run pytest` at all. I suspect you had `pytest` installed globally somehow.

The `pytest.mark` usage is from `pytest-datafiles` as in [this example](https://github.com/omarkohl/pytest-datafiles/blob/master/examples/example_2.py). Perhaps you didn't have `pytest-datafiles` installed and that is where the error came from.

I added the dev requirements to the `pyproject.toml` in a way uv can find them. If you could see if the current develop branch works better for you, I would be grateful.
