### 2025-03-07 66ba84e
tests for `run.py` and `core.py`
