+++ pyproject.toml
@@ -43,6 +43,7 @@ dev = [
     "mypy",
     "pytest>=8.3.2",
     "pytest-cov",
+    "pytest-mock",
     "pytest-datafiles",
     "griffe==1.2.0"
 ]
@@ -58,6 +59,7 @@ dev-dependencies = [
     "mypy",
     "pytest>=8.3.2",
     "pytest-cov",
+    "pytest-mock",
     "pytest-datafiles",
     "griffe==1.2.0"
 ]
