### 2025-03-07 d709887
Improve documentation and fix some type errors
