### 2025-03-14 c618727
filter metadata table patch
