### 2025-03-14 73c9995
some typing fixes
