### 2025-03-16 #7870 @xoolive
Hi all

My comments are pretty similar to those by the other reviewers.

Some items cannot be checked for obvious reasons, e.g. community guidelines, automated tests, etc.

Paper:
- I think you should clarify what you mean by data reduction. I understand it as "computing metrics from raw data" but cannot say that for sure.

Repository:
- No automated tests (github actions or so)
- The tests are also very light, only very basic functionalities seem tested
- No installation instructions
- I did `uv venv` to create an environment and tested the example in the documentation, but it does not work (after creating a venv):

```sh
pydre/project_files on  master via 🐍 3.13.2 via pydre
➜ uv run python -m pydre.run -p tutorial_project.toml -d Experimenter_S1_Tutorial_11002233.dat -o tutorial_out.csv
      Built pydre @ file:///Users/xo/pydre
Installed 12 packages in 144ms
0it [00:00, ?it/s]
2025-03-16 14:55:19.847 | ERROR    | pydre.project:processDatafiles:261 - No results found; no metrics data generated
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "/Users/xo/pydre/src/pydre/run.py", line 51, in <module>
    main()
    ~~~~^^
  File "/Users/xo/pydre/src/pydre/run.py", line 46, in main
    p.processDatafiles(numThreads=12)
    ~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^
  File "/Users/xo/pydre/src/pydre/project.py", line 262, in processDatafiles
    result_dataframe = pl.from_dicts(results_list)
  File "/Users/xo/pydre/.venv/lib/python3.13/site-packages/polars/convert/general.py", line 209, in from_dicts
    raise NoDataError(msg)
polars.exceptions.NoDataError: no data, cannot infer schema
```

Please ping me here again when I can have a second look.
