+++ tests/test_data/good_projectfiles/test_time_roi_2.toml
@@ -1,0 +1,6 @@
+[rois.roi2]
+type = "time"
+filename = "C:\\Users\\aatchley\\Documents\\test_time_2.csv"
+
+[metrics.mean]
+function = "colMean"
