+++ tests/test_data/good_projectfiles/test_time_roi_1.toml
@@ -1,0 +1,7 @@
+[rois.roi1]
+type = "time"
+filename = "C:\\Users\\aatchley\\Documents\\test_time_1.csv"
+
+[metrics.meanVelocityOver20]
+function = "meanVelocity"
+cutoff = 20
