+++ tests/test_data/good_projectfiles/test_time_roi_1.toml
@@ -3,5 +3,5 @@ type = "time"
 filename = "C:\\Users\\aatchley\\Documents\\test_time_1.csv"
 
 [metrics.meanVelocityOver20]
-function = "meanVelocity"
-cutoff = 20
+function = "colMean"
+var = "DatTime"
