+++ tests/test_data/good_projectfiles/test_time_roi_2.toml
@@ -4,3 +4,4 @@ filename = "C:\\Users\\aatchley\\Documents\\test_time_2.csv"
 
 [metrics.mean]
 function = "colMean"
+var = "DatTime"
