+++ src/pydre/rois.py
@@ -18,7 +18,7 @@ class ROIProcessor(object, metaclass=ABCMeta):
     @abstractmethod
     def split(
         self, sourcedrivedata: pydre.core.DriveData
-    ) -> list[pydre.core.DriveData]:
+    ) -> Iterable[pydre.core.DriveData]:
         """Splits the drivedata object according to the ROI specifications.
 
         Parameters:
@@ -91,25 +91,6 @@ class TimeROI(ROIProcessor):
         the 'roi' field of the objects will be filled with the roi tag listed
         in the roi definition file column name
         """
-        # output_list = []
-        #
-        # if sourcedrivedata.PartID in self.rois.keys():
-        #     for roi, duration in self.rois[sourcedrivedata.PartID]:
-        #         start, end = duration
-        #         new_data = sliceByTime(start, end, timecol, sourcedrivedata.data)
-        #         new_ddata = pydre.core.DriveData(sourcedrivedata, new_data)
-        #         new_ddata.roi = roi
-        #         output_list.append(new_ddata)
-        # return output_list
-        #
-        # if sourcedrivedata.PartID in self.rois.keys():
-        #     for roi, duration in self.rois[sourcedrivedata.PartID]:
-        #         start, end = duration
-        #         new_data = sliceByTime(start, end, timecol, sourcedrivedata.data)
-        #         new_ddata = pydre.core.DriveData(sourcedrivedata, new_data)
-        #         new_ddata.roi = roi
-        #         output_list.append(new_ddata)
-        # return output_list
         output_list = []
         matching_rois = self.rois.copy()
         if len(self.rois_meta) > 0:
@@ -146,20 +127,6 @@ class TimeROI(ROIProcessor):
         # time1-time2 where time1 or time 2 are either hr:min:sec or min:sec
         # example:  1:15:10-1:20:30
         # example : 02:32-08:45
-
-        # pair_regex = r"([\d:])-([\d:])"
-        # time_regex = r"(?:(\d+):)?(\d+):(\d+)"
-        # pair_result = re.match(pair_regex, duration)
-        # first_time_str, second_time_str = pair_result.group(1, 2)
-        # first_time_result = re.match(time_regex, first_time_str)
-        # second_time_result = re.match(time_regex, second_time_str)
-        # first_time = first_time_result.group(2) * 60 + first_time_result.group(3)
-        # if first_time_result.group(1):
-        #     first_time += first_time_result.group(1) * 60 * 60
-        # second_time = second_time_result.group(2) * 60 + second_time_result.group(3)
-        # if second_time_result.group(1):
-        #     second_time += second_time_result.group(1) * 60 * 60
-        # return (first_time, second_time)
 
         regex = r"(?:(\d{1,2}):)?(\d{1,2}):(\d{2})"
         pair_result = re.match(regex, duration)
@@ -199,7 +166,7 @@ class SpaceROI(ROIProcessor):
 
     def split(
         self, sourcedrivedata: pydre.core.DriveData
-    ) -> list[pydre.core.DriveData]:
+    ) -> Iterable[pydre.core.DriveData]:
         return_list: list[pydre.core.DriveData] = []
 
         for roi_name, roi_location in self.roi_info.items():
@@ -225,11 +192,6 @@ class SpaceROI(ROIProcessor):
             )
 
             if region_data.height == 0:
-                # try out PartID to get this to run cgw 5/20/2022
-                # logger.warning("No data for SubjectID: {}, Source: {},  ROI: {}".format(
-                #    ddata.SubjectID,
-                #    ddata.sourcefilename,
-                #    self.roi_info.roi[i]))
                 logger.warning(
                     "No data for SubjectID: {}, Source: {},  ROI: {}".format(
[…]
