+++ docs/index.md
@@ -1,13 +1,13 @@
 
 # Pydre Overview
 
-Pydre is a software package designed to analyze driving data taken from driving simulation research studies. Specifically, it was built around the data file format generated by [SimObserver](https://www.faac.com/realtime-technologies/products/simobserver-pro/), which is software sold by Realtime Technologies for recording data from their SimCreator simulation system.
+Pydre is a software package designed to reduce the large quantity of  driving data taken from driving simulation research studies into a smaller set of relevant metrics that can be more easily used for statistical analysis.
 
 ## Motivation
 
 Driving simulation data can not be easily processed without preprocessing. This preprocessing is often similar between experiments but not identical. Experimenters often need to write data processing code bespoke for each experiment, but this is unwieldy and does not lend itself to reproducible science.
 
-Although Pydre was developed specifically to process the datafiles output from SimObserver, it could be modified easily to process any simulator output that has one row per timestep.
+Although Pydre was developed specifically to process the data in a format (`.dat` files which are TSV) that has been created by [SimObserver](https://www.faac.com/realtime-technologies/products/simobserver-pro/), it could be modified easily to process any simulator output that has one row per timestep. Since the input data is a flat file per drive, not much work would be needed to read data formatted in a similar way.
 
 SimObserver collects information during the simulation drive, such as vehicle speed and following distance.  This information is collected every 1/60th of a second, which results in tens of thousands of rows for a study session. Experimenters are often interested in specific metrics that look at trends, averages or other quantities that are determined by looking at longer periods of time. Processing a large experimental dataset into a smaller file of quantified metrics is what Pydre was designed to do. 
 
@@ -17,7 +17,7 @@ Pydre takes a drive data file with data that is difficult to interpret in raw fo
 
 # Structure
 
-Pydre works by reading in `.dat` files (creating DriveData objects), then applying filters, region of interest splitters and metrics to each one.
+Pydre works by reading in a sequence of raw data files from a driving simulator output (creating DriveData objects), then applying filters, region of interest splitters and metrics to each one. The output is a csv file with the results of the metrics for each region of interest (ROI) in the data.
 
 ![](images/pydre%20data%20pipeline.png "Data Pipeline")
 
