### 2025-03-17 160702d
move custom metrics to tutorial
