### 2025-03-18 d33340f
use polars specific assert in test
