### 2025-03-18 e55e164
Tightened up documetation
ran reformatter on code
