### 2025-03-18 a263065
Better custom metrics setup documentation
