### 2025-03-19 0322451
added tests for parseTimestamp
