+++ tests/test_data/good_projectfiles/test_time_roi_2.toml
@@ -1,6 +1,6 @@
 [rois.roi2]
 type = "time"
-filename = "C:\\Users\\aatchley\\Documents\\test_time_2.csv"
+filename = "../test_datfiles/Experimenter_S1_Tutorial_11002233.dat"
 
 [metrics.mean]
 function = "colMean"
