### 2025-03-19 #7870 @tkerwin
@fredshone @xoolive @fxjung

Thank you all for your comments. I have released an updated version (25.1) on pypi and have even more up-to-date documentation on the master branch. This includes:

- Support for using *uv*
- Updated installation and usage documentation
- An example project directory.
- Automated testing through github actions
- Added contribution guidelines

I have also added clarification in the paper itself about data reduction:

> In this context, we describe the process of converting raw data to a meaningful set of discrete metrics as “data reduction".

I have added some new tests and continue to add them. I do realize that our overall test coverage is low. However, there are quite a few metrics and filters that are in the repository currently that were written for analysis of specific research studies. These functions are not yet fully documented or intended to be used for end users at this time.

I would appreciate you pulling the changes and taking another look at the documentation. Please let me know of any other issues that you see.

Thanks again.
