### 2025-03-20 8a010e9
Better error message rather than throw exceptions
