### 2025-03-20 06bbf8b
Refactor R2D specific functions
