+++ src/pydre/metrics/common.py
@@ -1,6 +1,5 @@
 from typing import Optional
 
-import re
 import polars as pl
 from polars import exceptions
 import pydre.core
@@ -614,69 +613,6 @@ def steeringReversalRate(drivedata: pydre.core.DriveData) -> float:
     return reversal_rate
 
 
-@registerMetric()
-def throttleReactionTime(drivedata: pydre.core.DriveData) -> Optional[float]:
-    """Calculates the time it takes to accelerate once follow car brakes (r2d)
-
-    Note: Requires data columns
-        - SimTime: Simulation time in seconds
-        - FollowCarBraking Status: Whether the follow car is braking
-        - LonAccel: Longitude acceleration
-        - Brake: Whether the ownship is braking
-
-    Returns:
-        Time in seconds from when the follow car braked to when the ownship started accelerating forward.
-    """
-
-    required_col = ["FollowCarBrakingStatus", "LonAccel", "SimTime", "Brake"]
-
-    try:
-        drivedata.checkColumnsNumeric(required_col)
-    except pl.exceptions.PolarsError:
-        return None
-
-    df = drivedata.data.select(
-        [
-            pl.col("FollowCarBrakingStatus"),
-            pl.col("SimTime"),
-            pl.col("LonAccel"),
-            pl.col("Brake"),
-            pl.col("Throttle"),
-        ]
-    )
-
-    if df.height < 1:
-        return None
-
-    initial_time = df.get_column("SimTime").item(0)
-
-    try:
-        df = df.filter(
-            pl.col("SimTime")
-            > df.filter(pl.col("Brake") > 3.0).get_column("SimTime").item(0)
-        )
-    except IndexError:
-        logger.warning(
-            f"No braking detected for roi {drivedata.roi} in file {drivedata.sourcefilename}"
-        )
-        return None
-
-    df_after_brake = df.filter(pl.col("Brake") == 0)
-
-    try:
-        time_of_accel = (
-            df_after_brake.filter(pl.col("LonAccel") > 0).get_column("SimTime").item(0)
-        )
-    except IndexError:
-        logger.warning(
-            f"No subsequent acceleration detected for roi {drivedata.roi} in file {drivedata.sourcefilename}"
-        )
-        return None
-
-    throttle_reaction_time = time_of_accel - initial_time
-    return throttle_reaction_time
-
-
 # mean dip time
 # number of dips
 # median dip time
@@ -720,122 +656,6 @@ def biopticDipMeasure(drivedata: pydre.core.DriveData):
         numDips,
         medianDipTime,
     ]
-
-
-@registerMetric()
-def eventSpeedRecoveryTime(
-    drivedata: pydre.core.DriveData, op_speed=15.64, tolerance=4
-) -> Optional[float]:
-    """Calculates the time it takes to accelerate back up to 'operational speed' (op_speed)
-    tolerance allows for proper classification of subject behavior variation
-    (defaults parameterized for R2D study processing)
-
-    Note: Requires data columns
-        - SimTime: Simulation time in seconds
-        - Velocity: m/s directional speed of Subject
-
-    Returns:
-        Time in seconds from the start of the CE to when the Subject returned to (operational speed - tolerance)
-    """
-
-    required_col = ["SimTime", "Velocity"]
-    lower_bound = op_speed - tolerance
-
-    try:
-        drivedata.checkColumnsNumeric(required_col)
-    except pl.exceptions.PolarsError:
-        return None
-
-    df = drivedata.data.select([pl.col("Velocity"), pl.col("SimTime"), pl.col("Brake")])
-
-    # filter df, so we take recovery time after initial slow-down
-    try:
-        df = df.filter(
-            pl.col("SimTime")
-            > df.filter(pl.col("Velocity") < lower_bound).get_column("SimTime").item(0)
-        )
[…]
