+++ src/pydre/metrics/R2DMetrics.py
@@ -1,0 +1,296 @@
+import re
+from typing import Optional
+
+import polars as pl
+from loguru import logger
+
+import pydre.core
+from pydre.core import ColumnsMatchError
+from pydre.metrics import registerMetric
+
+
+@registerMetric(
+    "R2DIDColumns", ["ParticipantID", "MatchID", "Case", "Location", "Gender", "Week"]
+)
+def R2DIDColumns(drivedata: pydre.core.DriveData):
+    ident = drivedata.metadata["ParticipantID"]
+    ident_groups = re.match(r"(\d)(\d)(\d)(\d\d\d\d)[wW](\d)", ident)
+    if ident_groups is None:
+        logger.warning("Could not parse R2D ID " + ident)
+        return [None, None, None, None, None, None]
+    participant_id = (
+        ident_groups.group(1)
+        + ident_groups.group(2)
+        + ident_groups.group(3)
+        + ident_groups.group(4)
+    )
+    case = ident_groups.group(1)
+    if case == "3":
+        case = "Case"
+    elif case == "5":
+        case = "Control"
+    elif case == "7":
+        case = "Enrolled"
+    location = ident_groups.group(2)
+    if location == "1":
+        location = "UAB"
+    elif location == "2":
+        location = "OSU"
+    elif location == "3":
+        location = "UA"
+    gender = ident_groups.group(3)
+    if gender == "1":
+        gender = "Male"
+    elif gender == "2":
+        gender = "Female"
+    match_id = ident_groups.group(4)
+    week = ident_groups.group(5)
+    return participant_id, match_id, case, location, gender, week
+
+
+@registerMetric()
+def throttleReactionTime(drivedata: pydre.core.DriveData) -> Optional[float]:
+    """Calculates the time it takes to accelerate once follow car brakes (r2d)
+
+    Note: Requires data columns
+        - SimTime: Simulation time in seconds
+        - FollowCarBraking Status: Whether the follow car is braking
+        - LonAccel: Longitude acceleration
+        - Brake: Whether the ownship is braking
+
+    Returns:
+        Time in seconds from when the follow car braked to when the ownship started accelerating forward.
+    """
+
+    required_col = ["FollowCarBrakingStatus", "LonAccel", "SimTime", "Brake"]
+
+    try:
+        drivedata.checkColumnsNumeric(required_col)
+    except pl.exceptions.PolarsError:
+        return None
+
+    df = drivedata.data.select(
+        [
+            pl.col("FollowCarBrakingStatus"),
+            pl.col("SimTime"),
+            pl.col("LonAccel"),
+            pl.col("Brake"),
+            pl.col("Throttle"),
+        ]
+    )
+
+    if df.height < 1:
+        return None
+
+    initial_time = df.get_column("SimTime").item(0)
+
+    try:
+        df = df.filter(
+            pl.col("SimTime")
+            > df.filter(pl.col("Brake") > 3.0).get_column("SimTime").item(0)
+        )
+    except IndexError:
+        logger.warning(
+            f"No braking detected for roi {drivedata.roi} in file {drivedata.sourcefilename}"
+        )
+        return None
+
+    df_after_brake = df.filter(pl.col("Brake") == 0)
+
+    try:
+        time_of_accel = (
+            df_after_brake.filter(pl.col("LonAccel") > 0).get_column("SimTime").item(0)
+        )
+    except IndexError:
+        logger.warning(
+            f"No subsequent acceleration detected for roi {drivedata.roi} in file {drivedata.sourcefilename}"
+        )
+        return None
+
+    throttle_reaction_time = time_of_accel - initial_time
+    return throttle_reaction_time
+
+
[…]
