### 2025-03-21 #7870 @xoolive
Also, non-blocking comment, careful with the warning [here](https://github.com/OSUDSL/pydre/actions/runs/13952750534)

```
No file matched to [**/uv.lock,**/requirements*.txt]. The cache will never get invalidated. Make sure you have checked out the target repository and configured the cache-dependency-glob input correctly.
```
