### 2025-03-21 #7870 @xoolive
Only blocking comment left for me is

- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
