### 2025-03-24 544a26c
cleaning up R2D data run exceptions + data type errors. confirmed compat with data packager.
