+++ src/pydre/metrics/R2DMetrics.py
@@ -61,6 +61,7 @@ def throttleReactionTime(drivedata: pydre.core.DriveData) -> Optional[float]:
     Returns:
         Time in seconds from when the follow car braked to when the ownship started accelerating forward.
     """
+    drivedata.data = drivedata.data.with_columns(drivedata.data["SimTime"].cast(pl.Float64))
 
     required_col = ["FollowCarBrakingStatus", "LonAccel", "SimTime", "Brake"]
 
@@ -126,7 +127,7 @@ def eventSpeedRecoveryTime(
     Returns:
         Time in seconds from the start of the CE to when the Subject returned to (operational speed - tolerance)
     """
-
+    drivedata.data = drivedata.data.with_columns(drivedata.data["SimTime"].cast(pl.Float64))
     required_col = ["SimTime", "Velocity"]
     lower_bound = op_speed - tolerance
 
@@ -181,7 +182,7 @@ def eventRecenterRecoveryTime(
         Time in seconds from when the Subject left (lane_offset +- tolerance )
         to when the Subject returned to (lane_offset +- tolerance)
     """
-
+    drivedata.data = drivedata.data.with_columns(drivedata.data["SimTime"].cast(pl.Float64))
     required_col = ["EventName", "LaneOffset", "SimTime"]
 
     try:
@@ -232,6 +233,7 @@ def reactionCheckVarVal(
     drivedata: pydre.core.DriveData, var: str, val: float
 ) -> Optional[float]:
     required_col = [var, "SimTime"]
+    drivedata.data = drivedata.data.with_columns(drivedata.data["SimTime"].cast(pl.Float64))
     try:
         drivedata.checkColumnsNumeric(required_col)
     except ColumnsMatchError:
@@ -255,6 +257,7 @@ def reactionTimeEventTrueR2D(
     drivedata: pydre.core.DriveData, var1: str, var2: str, val1: float, val2: float
 ):
     required_col = [var1, var2, "SimTime"]
+    drivedata.data = drivedata.data.with_columns(drivedata.data["SimTime"].cast(pl.Float64))
     try:
         drivedata.checkColumnsNumeric(required_col)
     except ColumnsMatchError:
