### 2025-03-24 4ff13f1
cleaner nonevent reporting when ran in bulk
