### 2025-03-24 e1b5086
added more tests for roi creation
