### 2025-03-25 OSUDSL/pydre#1 @fredshone
sorry for slow response.

`uv run pytest` all good on develop and master.
