### 2025-03-25 #7870 @fredshone
All good. Nice job @tkerwin .

Some thoughts for the future health of your project, (not part of the review):

- Given your low coverage I think you should include coverage in the CI and fail under what you have covered now (37%),  you can just add fail under to pyproject.toml, [eg](https://github.com/arfc/osier/blob/main/pyproject.toml). Fail unders maintain a nudge towards good future behavior.
- I agree with @xoolive that normally expect to see a little more in the readme. When I'm looking for projects to use I judge them by the readme.
- I believe this project will have more uptake if you rename it dr.dre.
