### 2025-03-25 7ae5fee
Restore polars version of eye glance metrics
