+++ src/pydre/filters/R2DFilters.py
@@ -1,377 +1,0 @@
-import re
-import pathlib
-
-import polars as pl
-from loguru import logger
-
-import pydre.core
-from pydre.filters import registerFilter
-
-THISDIR = pathlib.Path(__file__).resolve().parent
-
-
-@registerFilter()
-def modifyCriticalEventsCol(drivedata: pydre.core.DriveData):
-    ident = drivedata.metadata["ParticipantID"]
-    ident_groups = re.match(r"(\d)(\d)(\d)(\d\d\d\d)[wW](\d)", ident)
-    if ident_groups is None:
-        logger.warning("Could not parse R2D ID " + ident)
-        return [None]
-    week = ident_groups.group(5)
-    scenario = drivedata.metadata["ScenarioName"]
-    if week == "1" and scenario == "Load, Event":
-        # between the x positions, change the critical event status to 1
-        drivedata.data = drivedata.data.with_columns(
-            pl.when(2165 < pl.col("XPos"), pl.col("XPos") < 2240)
-            .then(1)
-            .when(pl.col("CriticalEventStatus") == 1)
-            .then(1)
-            .otherwise(0)
-            .alias("CriticalEventStatus")
-        )
-    return drivedata
-
-
-@registerFilter()
-def ValidateDataStartEnd(drivedata: pydre.core.DriveData, dataFile="", tol=100, trim_data=False):
-    """
-    Ensure that the end of the drive data fits into the expected
-    format - by distance & time.
-
-    "early" - when xPos is less than expected
-    "late" - when xPos is greater than expected
-    "valid" - when all xPos fall within tolerance range
-
-    :arg: dataFile: source of truth for start/end - csv filepath
-    :arg: tol: tolerance of distance that permits "valid"
-    """
-    ident = drivedata.metadata["ParticipantID"]
-    scenario = drivedata.metadata["ScenarioName"]
-    # (control=5/case=3)(UAB=1/OSU=2)(Male=1/Female=2)(R2D_ID)w(Week Num)
-    ident_groups = re.match(r"(\d)(\d)(\d)(\d\d\d\d)[wW](\d)", ident)
-    if ident_groups is None:
-        logger.warning("Could not parse R2D ID " + ident)
-        return [None]
-    week = ident_groups.group(5)
-    df = drivedata.data
-
-    if dataFile != "":
-        merge_df = pl.read_csv(source=dataFile)
-        merge_df = merge_df.filter(
-            merge_df.get_column("Scenario") == scenario,
-            merge_df.get_column("Week") == week,
-        )
-    else:
-        logger.warning(f"Failed to read csv at {dataFile} - check local path.")
-        return [None]
-
-    if len(merge_df) > 0:
-        status = []
-        expected_start = merge_df["Start Pos"][0]
-        expected_end = merge_df["End Pos"][0]
-        actual_start = df["XPos"][0]
-        actual_end = df["XPos"][-1]
-
-        if actual_start is None:
-            logger.warning(
-                f"Start point in data is null for {ident}. See data for issue."
-            )
-
-        if actual_end is None:
-            logger.warning(
-                f"End point in data is null for {ident}. See data for issue."
-            )
-
-        if actual_start < expected_start - tol:
-            status.append("earlyStart")
-        elif actual_start > expected_start + tol:
-            status.append("lateStart")
-
-        if actual_end < expected_end - tol:
-            status.append("earlyEnd")
-        elif actual_end > expected_end + tol:
-            status.append("lateEnd")
-
-        if not status:
-            status.append("valid")
-
-        status_value = "&".join(status)
-        logger.debug(f"GOT STATUS VALUE: {status_value}")
-        df = df.with_columns(pl.lit(status_value).alias("validityCheck"))
-        if trim_data:
[…]
