### 2025-03-28 #7870 @fxjung
In "Pydre: A Python package for driving simulation data reduction", the author introduces Pydre, a Python data analysis pipeline used to filter and analyze data retrieved from physical driving simulators. The package seems useful and straightforward to apply. I have some minor comments regarding the installation process and instructions, the tutorial, and the manuscript, that I will list below.

I agree with the other reviewers that a slightly more detailed README on the GitHub repository, while not mandatory, would greatly improve accessibility.

## Installation

- The [installation instructions](https://osudsl.github.io/pydre/tutorial/installation/) focus on `rye`, which is superseded by `uv`. This is fine, however, I would greatly appreciate if the instructions using the standard tool (`pip`) were given as well. While the instructions do mention that Pydre may be installed via `pip`, the name of the package on PyPI is not specified.
- I used `pip` to set up the dev environment. However, `pip` does not understand `pyproject.toml`'s `dependency-groups` section. Simply renaming it to `project.optional-dependencies` would allow to set up the Pydre by `pip install -e '.[dev]'` (which is what I did, locally)

## Tutorial

- The output of the [getting started tutorial](https://osudsl.github.io/pydre/tutorial/getting_started/) is incorrect; the data file does not contain a variable `LanePosition` as specified in the example TOML file. This should probably read `LaneOffset` instead.
- Also, the tutorial instructs to store the TOML as `tutorial_project.toml`, whereas the call to `pydre.run` refers to `tutorial.toml`.
- Always writing `python -m pydre.run` is a bit cumbersome. The package does install a shell entrypoint, but it is undocumented and defective:
  ```bash
% pydre-run -p tutorial_project.toml -d Experimenter_S1_Tutorial_11002233.dat                                                                        :(
Traceback (most recent call last):
  File "/home/jung/.pyenv/versions/pydre/bin/pydre-run", line 8, in <module>
    sys.exit(run())
             ^^^^^
TypeError: 'module' object is not callable
```
  I also think that just "`pydre`" would be a more suited name for the entrypoint.
- Generally, it would be great if the tutorial would also specify a filter and an ROI.

## Manuscript

The manuscript is very well written, making it easy to understand what the package does, what its purpose is, and how it relates to other software. The data pipeline chart is helpful as well. I have a few minor comments regarding making the paper more accessible to outsiders.

- When reading the summary paragraph, it was not immediately clear to me that Pydre is about data obtained from *physical*, hardware-bound driving simulators that interact with human drivers, and not about "autonomous" computer simulations of driving behavior. Please make this a bit clearer to not confuse people outside the specific field of driving simulators.
- On P1L27 you mention "SimCreator" without introducing it or providing a reference.
- At the start of the architecture section (P2L48), it would be great to have one or two examples for the raw drive data ingested.
- In the metric functions section (P3L70) please add a short explanation on how to use Pydre's output---what are typical use cases?
- On P3L77, you mention "SimObserver"---please provide a reference.
