### 2025-03-28 abcaefb
pip compatibility
