+++ docs/tutorial/installation.md
@@ -68,6 +68,58 @@ If you encounter any issues:
 2. Verify that your [project file](../explanation/project_files.md) is properly formatted.
 3. If you are still having problems, please [open an issue on GitHub](https://github.com/OSUDSL/pydre/issues).
 
+# Setting up pydre with *pip*
+
+## 1. Install Python
+
+Ensure you have Python installed on your system. You can download it from [python.org](https://www.python.org/).
+
+## 2. Set up your project directory
+
+Create a new project directory and navigate to it:
+
+```bash
+mkdir my_pydre_project
+cd my_pydre_project
+```
+
+## 3. Create a virtual environment
+
+Set up a virtual environment to isolate your project dependencies:
+
+```bash
+python -m venv venv
+```
+
+Activate the virtual environment:
+
+- On Windows:
+  ```bash
+  venv\Scripts\activate
+  ```
+- On macOS/Linux:
+  ```bash
+  source venv/bin/activate
+  ```
+
+## 4. Install Pydre
+
+Use pip to install Pydre:
+
+```bash
+pip install pydre
+```
+
+### 5. Verify Installation
+
+Check that Pydre was installed correctly:
+
+```bash
+python -m pydre.run --help
+```
+
+You should see the help output showing available command line options.
+
 # Setting up a development environment
 
 If you want to contribute to Pydre or modify its source code, you can set up a development environment. You will need to clone the Pydre repository:
