### 2025-03-31 86e30c6
Update docs for new entry point.
