+++ joss/paper.bib
@@ -48,3 +48,15 @@
   eventtitle = {{{SoutheastCon}} 2020},
   isbn = {978-1-72816-861-6}
 }
+
+@software{SimCreator,
+  title = {{{SimCreator}}},
+  url = {https://www.faac.com/realtime-technologies/products/simcreator/},
+  organization = {Realtime Technologies}
+}
+
+@software{SimObserver,
+  title = {{{SimObserver}}},
+  url = {https://www.faac.com/realtime-technologies/products/simobserver-pro/},
+  organization = {Realtime Technologies}
+}
