+++ joss/paper.md
@@ -21,7 +21,7 @@ date: 21 January 2025
 
 Driving simulators are complex pieces of equipment that are extremely valuable in experimental studies on driving behavior. Simulation technology is very useful in experimentation when doing something in real life is either too dangerous or too expensive. In driving studies, especially for crash-imminent and driver distraction scenarios, the necessary experimental conditions are often both of these things.
 
-The Python package *Pydre* (pronounced pie-dray), described in this document, provides a cohesive framework for data reduction in the context of driving simulation.  This software package is intended to be broadly useful for all driving simulation researchers.  
+The Python package *Pydre* (pronounced pie-dray), described in this document, provides a cohesive framework for data reduction in the context of real-time, physical driving simulation systems that are used to investigate driving behavior. This software package is intended to be broadly useful for all driving simulation researchers that process data from such mixed and virtual reality driving simulators. 
 
 # Statement of need
 
@@ -29,7 +29,7 @@ The driving simulators commonly used for the investigation of driving behavior g
 
 Reducing the raw data from driving simulation experiments has been a task required for the statistical analysis of those experiments for some time [@michellelreyesSimulatorData2011]. Many different individual metrics have been developed over the years as variables of investigatory interest [@mcmanusPrincipalComponents2024]. 
 
-Other researchers have proposed data reduction software packages. Loab et al. describe a Matlab toolkit that performs data reduction [@loebSimulatedDriving2014]. The notion of "blocks" in that work is directly tied to the architecture of a previous version of SimCreator, making it less universal. Pawar et al. describe a framework of data reduction similar to Pydre, with regions of interest and metrics [@pawarDataReduction2020]. However, it lacks the extensible filter system. Neither of these packages are available online in a ready-to-use, runnable form, and both are described in a way that focuses on a specific experimental scenario. 
+Other researchers have proposed data reduction software packages. Loab et al. describe a Matlab toolkit that performs data reduction [@loebSimulatedDriving2014]. The notion of "blocks" in that work is directly tied to the architecture of a previous version of SimCreator [@SimCreator], making it less universal. Pawar et al. describe a framework of data reduction similar to Pydre, with regions of interest and metrics [@pawarDataReduction2020]. However, it lacks the extensible filter system. Neither of these packages are available online in a ready-to-use, runnable form, and both are described in a way that focuses on a specific experimental scenario. 
 
 This data reduction is a necessary component of the most common type of driving-data based analysis of participant simulator performance. This work could be accomplished by bespoke processing scripts for a particular project. However, these one-off scripts are often not taken seriously as reusable code and can lead to fragile software projects that can not be easily shared between researchers, even at the same lab. In addition, many approaches using multi-stage processing with multiple programs can lead to intermediate data files and data pipelines that can lead to running code on the incorrect "stage" of the data.
 
@@ -41,6 +41,10 @@ In using Pydre to perform data reduction, researchers write a project file descr
 
 ![Pydre data pipeline. Data files are filtered individually, broken up in to regions of interest, then metrics are processed for each region. \label{fig:data_flow}](pydre data pipeline.png)
 
+
+## Raw data files
+
+The raw data files are the output of the driving simulator. They are typically in a tabular format, such as CSV or TSV, and contain time-series data for each participant in the study. Each row in the data file represents a single time step in the simulation, and each column represents a different variable, such as vehicle speed, steering wheel angle, or roadway position, as exported. Each individual data file contains all the data for a single participant driving a single scenario.
 
 ## Data filters
 
@@ -55,11 +59,14 @@ Regions of interest (ROIs) are partitions of the driving data. This is especiall
 
 ## Metric functions
 
[…]
