+++ tests/test_data/good_projectfiles/test_ignore_1.toml
@@ -1,0 +1,10 @@
+[config]
+datafiles = ["test_data\\test_datfiles\\CrossCorrTest_Sub_1_Drive_*.dat"]
+
+[rois.roi2]
+type = "time"
+filename = "tests/test_data/csv/test_ignore.csv"
+
+[metrics.mean]
+function = "colMean"
+var = "DatTime"
