+++ project_files/test_time_roi_2.toml
@@ -1,9 +1,11 @@
 [config]
 datafiles = ["C:\\Users\\aatchley\\Documents\\test_folder_2\\*.dat"]
 ignore = ["C:\\Users\\aatchley\\Documents\\test_folder_2\\Experimenter_2_Test B_1734106849.dat", "C:\\Users\\aatchley\\Documents\\test_folder_2\\Experimenter_3_Accommodation Drive_1734110222.dat"]
+
 [rois.roi2]
 type = "time"
 filename = "C:\\Users\\aatchley\\Documents\\test_time_2.csv"
+
 [metrics.mean]
 function = "colMean"
 var = "DatTime"
