### 2025-04-03 #7870 @arfon
Thanks @tkerwin! Given the strong positive feedback thus far from reviewers, I believe we're very close to being able to accept and publish. If @fredshone, @xoolive, and @fxjung have a chance to return for a quick :+1: to confirm, that would be great, but if I don't hear back in the next week we can proceed regardless.

Also, @fredshone, I feel duty-bound to celebrate review humour, so thank you for this one 😆

> I believe this project will have more uptake if you rename it dr.dre.
