### 2025-04-03 #7870 @fxjung
@tkerwin Thanks for the comprehensive updates, everything looks good now and works fine. I also think the paper is now much easier to understand, especially for outsiders.

Two small things:

- In the manuscript, P3L95, "Pydre was built to ingest data formatted like the output SimObserver [...]"---there is probably an "of" missing after "output"
- In the [getting started tutorial](https://osudsl.github.io/pydre/tutorial/getting_started/), the first call to pydre uses backslashes in the path (Windows) and calls `pydre`, whereas the second forward slashes (Unix) and calls `rye run pydre`.

Other than that, I think it's good to go!
