### 2025-04-03 8f6ba65
Fix strange behavior when binary block  column starts on 1
