### 2025-04-09 #7870 @tkerwin
I've made a new release of the software (25.2.1) on github and pypi.
I uploaded an archive of the 25.2.1 release on figshare:  https://doi.org/10.6084/m9.figshare.28765025.v1
This should be the final draft of the article:

Please let me know if there is anything remaining for the submission.
