### 2025-04-10 ba46f4b
trimPreAndPostDrive function
