### 2025-04-14 4be6d86
redo ignore list
ignore list now removes matching base filenames
