### 2025-04-18 4f7fad5
docstrings edits
