### 2025-04-18 4b09081
Minor changes to nullifyOutlier
- Added column check
- Added test functions
