+++ tests/test_filters_common.py
@@ -1,9 +1,12 @@
 import polars as pl
+import polars.datatypes
 import pytest
 import pydre.core
 import polars.testing
 
 from pydre.filters.common import trimPreAndPostDrive
+from pydre.filters.common import nullifyOutlier
+
 
 
 def test_trim_pre_and_post_drive():
@@ -102,3 +105,121 @@ def test_trim_pre_and_post_drive_missing_column():
     with pytest.raises(Exception):
         trimPreAndPostDrive(dd)
 
+
+def test_nullify_outlier_default_parameters():
+    # Create a sample Polars DataFrame with outliers in HeadwayDistance
+    data = {
+        "SimTime": [0, 1, 2, 3, 4, 5],
+        "HeadwayDistance": [500.0, 800.0, 1500.0, 2000.0, 950.0, 700.0],
+    }
+    df = pl.DataFrame(data)
+    dd = pydre.core.DriveData.init_test(df, "test.dat")
+
+    # Apply the filter with default parameters (HeadwayDistance, threshold=1000)
+    filtered_dd = nullifyOutlier(dd)
+
+    # Expected result (values >1000 replaced with None)
+    expected_df = pl.DataFrame({
+        "SimTime": [0, 1, 2, 3, 4, 5],
+        "HeadwayDistance": [500.0, 800.0, None, None, 950.0, 700.0],
+    })
+
+    pl.testing.assert_frame_equal(filtered_dd.data, expected_df)
+
+
+def test_nullify_outlier_custom_threshold():
+    # Create a sample Polars DataFrame with outliers
+    data = {
+        "SimTime": [0, 1, 2, 3, 4],
+        "HeadwayDistance": [50.0, 150.0, 200.0, 500.0, 100.0],
+    }
+    df = pl.DataFrame(data)
+    dd = pydre.core.DriveData.init_test(df, "test.dat")
+
+    # Apply the filter with custom threshold
+    filtered_dd = nullifyOutlier(dd, threshold=200)
+
+    # Expected result (values >200 replaced with None)
+    expected_df = pl.DataFrame({
+        "SimTime": [0, 1, 2, 3, 4],
+        "HeadwayDistance": [50.0, 150.0, 200.0, None, 100.0],
+    })
+
+    pl.testing.assert_frame_equal(filtered_dd.data, expected_df)
+
+
+def test_nullify_outlier_custom_column():
+    # Create a sample Polars DataFrame with outliers in a different column
+    data = {
+        "SimTime": [0, 1, 2, 3, 4],
+        "Speed": [30.0, 120.0, 150.0, 200.0, 50.0],
+    }
+    df = pl.DataFrame(data)
+    dd = pydre.core.DriveData.init_test(df, "test.dat")
+
+    # Apply the filter with custom column name
+    filtered_dd = nullifyOutlier(dd, col="Speed", threshold=100)
+
+    # Expected result (values >100 replaced with None)
+    expected_df = pl.DataFrame({
+        "SimTime": [0, 1, 2, 3, 4],
+        "Speed": [30.0, None, None, None, 50.0],
+    })
+
+    pl.testing.assert_frame_equal(filtered_dd.data, expected_df)
+
+
+def test_nullify_outlier_no_outliers():
+    # Create a sample Polars DataFrame with no outliers
+    data = {
+        "SimTime": [0, 1, 2, 3, 4],
+        "HeadwayDistance": [500.0, 600.0, 700.0, 800.0, 900.0],
+    }
+    df = pl.DataFrame(data)
+    dd = pydre.core.DriveData.init_test(df, "test.dat")
+
+    # Apply the filter
+    filtered_dd = nullifyOutlier(dd, threshold=1000)
+
+    # No values should change
[…]
