### 2025-04-21 #7870 @arfon
@fredshone, @xoolive, @fxjung – many thanks for your reviews here! JOSS relies upon the volunteer effort of people like you and we simply wouldn't be able to do this without you ✨
